Create ClamAV_DirectoryScan.ps1
Basic PowerShell "Scriptlet" that allows for the quick scan of any given directory with Clam AV. This Script assumes that you have installed ClamAV on your C: Drive in Program Files, otherwise it will need negligible modification. Line six, as of this commit, (cd "C:\Program Files\ClamAV")

Furthermore this script will prompt the user for;

Directory to scan
Directory to store quarantined files
Directory to store the log file(automatically appends the log)

Lastly, the script will print out the log and stay alive for 2 minutes, another negligible modification.


In order to run this script you may need to changer the file permissions for the downloaded file, as well as Unblock this script via PowerShell which will obviously need to be run as Administrator. I am in now way responsible for the loss of any files, and furthermore you'll likely need to run the command below, note, it assumes you've put the file on your desktop and that your username is username... Be a good sport change the Path to reflect where ever you have downloaded this file to and ensure to include the filename.

> Unblock-File -Path C:\Users\Username\Desktop\ClamAV_DirectoryScan.ps1.ps1